+++ src/app/menu/layout.tsx
@@ -1,6 +1,6 @@
-import MainLayout from "@/components/layout/MainLayout";
+import ClientLayout from "@/components/layout/ClientLayout";
 import { TypeChildren } from "@/types/common";
 
 export default function Layout({ children }: TypeChildren) {
-  return <MainLayout>{children}</MainLayout>;
+  return <ClientLayout>{children}</ClientLayout>;
 }
