+++ src/app/store/_components/modals/StoreApplicationModal.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+/* eslint-disable react-hooks/exhaustive-deps */
 import ModalWithTitle from "@/components/modal/largeModalLayout";
 import { ChangeEvent, useEffect, useState } from "react";
 import { FormProvider, useForm } from "react-hook-form";
