+++ src/app/store/info/page.tsx
@@ -60,7 +60,7 @@ export default function StoreInfo() {
 
   useEffect(() => {
     setActiveMenu("매장 정보");
-  }, []);
+  }, [setActiveMenu]);
 
   function TableRow({
     children,
@@ -73,8 +73,8 @@ export default function StoreInfo() {
     );
   }
 
-  const handleDeleteItem = (item: OriginItem) => {
-    console.log(item);
+  const handleDeleteItem = () => {
+    // delete item: OriginItem
   };
 
   // const handle
@@ -93,8 +93,8 @@ export default function StoreInfo() {
     setCountryOfOrigins(newOrigins);
   };
 
-  const submitHandler = (data: TypeStoreInfo) => {
-    console.log(data);
+  const submitHandler = () => {
+    // submit data: TypeStoreinfo
   };
 
   return (
@@ -202,7 +202,7 @@ export default function StoreInfo() {
                                 ? "mr-1/2 flex w-full items-center justify-center md:px-2 lg:px-4"
                                 : "flex w-full items-center justify-center"
                             }
-                            onClick={() => handleDeleteItem(item)}
+                            onClick={handleDeleteItem}
                           >
                             <DeleteIcon
                               color="#F22020"
